feat: change harmony com some quest
