+++ packages/taro-harmony/src/components/components-harmony/navigator/index.js
@@ -1,12 +1,9 @@
 import router from '@system.router'
-import { createOption } from '../utils'
+import { createOption, queryToJson } from '../utils'
 
 export default createOption({
   props: {
     id: {
-      default: ''
-    },
-    cls: {
       default: ''
     },
     openType: {
@@ -55,7 +52,7 @@ export default createOption({
     // switchTab目前实现不了
     const methodName = this.targetObj[method]
     const [uri, queryString = ''] = url.split('?')
-    const params = this.queryToJson(queryString)
+    const params = queryToJson(queryString)
 
     const uriObj = url ? { uri: uri.replace(/^\//, '') } : {}
 
@@ -69,37 +66,6 @@ export default createOption({
       ...uriObj,
       ...paramsObj
     })
-  },
-
-  queryToJson (str) {
-    const dec = decodeURIComponent
-    const qp = str.split('&')
-    const ret = {}
-    let name
-    let val
-    for (let i = 0, l = qp.length, item; i < l; ++i) {
-      item = qp[i]
-      if (item.length) {
-        const s = item.indexOf('=')
-        if (s < 0) {
-          name = dec(item)
-          val = ''
-        } else {
-          name = dec(item.slice(0, s))
-          val = dec(item.slice(s + 1))
-        }
-        if (typeof ret[name] === 'string') {
-          ret[name] = [ret[name]]
-        }
-
-        if (Array.isArray(ret[name])) {
-          ret[name].push(val)
-        } else {
-          ret[name] = val
-        }
-      }
-    }
-    return ret
   },
 
   onTouchStart () {
